backend_django/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.0
numpy==1.24.3
pandas==2.1.1
scikit-learn==1.3.2
xgboost==2.0.1
geopy==2.4.0
requests==2.31.0
python-dotenv==1.0.0
gunicorn==21.2.0
psycopg2-binary==2.9.9
django-environ==0.11.2
httpx==0.25.0